+++ config/routes.rb
@@ -5,7 +5,15 @@ Rails.application.routes.draw do
   root to: "infrastructures#index"
 
   resources :infrastructures, only: [:show, :index] do
-    resources :environment_variables, only: [:new, :create, :destroy, :index]
+    resources :environment_variables, only: [:new, :create, :destroy, :index] do
+      collection do
+        resources :env_files, only: [:new, :create] do
+          collection do
+            post "confirm", to: "env_files#confirm"
+          end
+        end
+      end
+    end
     resources :infrastructure_variables, only: [:new, :create, :destroy, :index], as: :variables
   end
 end