+++ CHANGELOG.md
@@ -21,6 +21,7 @@ The format is based on [Keep a Changelog 1.0.0].
 - secret environment variables are hidden by default
 - secret environment variables can be shown and hidden individually by clicking the cell
 - infrastructure variables are shown for every infrastructure (not just those running on the core AWS account)
+- multiple environment variables can be added at once with a .env file
 
 [unreleased]: TODO
 [keep a changelog 1.0.0]: https://keepachangelog.com/en/1.0.0/